Revised test cases for RUST main
