+++ prata/src/main.rs
@@ -20,7 +20,18 @@ fn main() {
  */
 fn run_tests()
 {
-    
+    //Unit tests for Channel
+    basic_test();
+    thread::sleep(time::Duration::from_millis(200));
+    remove_ip_test();
+    thread::sleep(time::Duration::from_millis(200));
+    add_data_test();
+    thread::sleep(time::Duration::from_millis(200));
+    remove_data_test();
+    thread::sleep(time::Duration::from_millis(200));
+    add_ip_ports_test();
+
+    //Integration tests
     test_fifo();
     thread::sleep(time::Duration::from_millis(200));
     test_custom_fifo();
@@ -31,13 +42,14 @@ fn run_tests()
     thread::sleep(time::Duration::from_millis(200));
     test_invalid_host_ip();
     thread::sleep(time::Duration::from_millis(200));
+
+
     
     //Error when publishing to channel, deleting channel then publishing
     //to the same channel. Error occurs from hanging TCP socket as it never
     //gets a reply from the deleted channel's original ip and port.
     //test_delete_channel_after_pub();
     
-    //test_set_port_range()
 }
 
 /**
@@ -63,7 +75,7 @@ fn test_set_port_range()
 fn test_invalid_host_ip()
 {
     println!("TEST: invalid_host_ip");
-    println!("---TEST: Expecting to see host fail error, If it prints, this tests passes. Consider boolean instead of print?");
+    println!("---TEST: Expecting to see host fail error. TEST PASSES if Invalid IP error occurs!");
     let m = master::Master {ipAddress: "207.168.0.122".to_string(), port: 25565, threading: true};
     m.host();
     thread::sleep(time::Duration::from_millis(200));
@@ -202,4 +214,98 @@ fn test_FIFO_and_Broadcast()
 
     m.terminate();
     println!("TEST: FIFO_and_Broadcast END");
+}
+fn add_ip_ports_test()
+{
+    println!("TEST: Channel - Add IP and Ports");
+    let config = channel::ChannelConfiguration::new(IPADDRESS.to_string(), 25565, "CUSTOM".to_string(), "fifo".to_string(), 20);
+    let mut c = channel::Channel::new(config);
+    c.add("192.168.0.0".to_string());
+    let ports = c.getPorts("192.168.0.0".to_string());
+
+    assert!(ports.fullRange == true, "TEST: Added single port to list, full range was not set")
+    
+}
+fn remove_data_test()
+{
+    println!("TEST: Channel - Remove Data");
+    let config = channel::ChannelConfiguration::new(IPADDRESS.to_string(), 25565, "CUSTOM".to_string(), "fifo".to_string(), 20);
+    let mut c = channel::Channel::new(config);
+
+    let data = c.getData();
+    assert!(data == "", "TEST: Constructs with data already inside FAIL");
+
+}
+fn add_data_test()
+{
+    println!("TEST: Channel - Add Data");
+    let config = channel::ChannelConfiguration::new(IPADDRESS.to_string(), 25565, "CUSTOM".to_string(), "fifo".to_string(), 20);
+    let mut c = channel::Channel::new(config);
+    c.addData("data element 1".to_string());
+    c.addData("data element 2".to_string());
+    c.addData("data element 3".to_string());
+    c.addData("data element 4".to_string());
+
+    let j = c.getData();
+    assert!(j == "data element 1".to_string(), "TEST: get data fail");
+
+    c.addData("the big\nhouse".to_string());
+    c.addData("\"what\" are you doing".to_string());
+    c.addData("\t\t\r\n".to_string());
+    c.addData("2 9 {{}}".to_string());
+
+    assert!(c.getData() == "data element 2", "TEST: get data 2 fail")
+}
+fn basic_test()
+{
+    println!("TEST: Channel - Basic");
+    let config = channel::ChannelConfiguration::new(IPADDRESS.to_string(), 25565, "CUSTOM".to_string(), "fifo".to_string(), 20);
+    let mut c = channel::Channel::new(config);
+    let listed = c.getListed();
+    assert!(listed.len() == 0, "TEST: Default constructor starts with ip listed");
+
+    c.add("192.168.0.0".to_string());
+    let listed = c.getListed();
+    assert!(listed.len() == 1, "TEST: Adding does not increase size by 1");
+
+    c.add("192.168.0.0".to_string());
+    let listed = c.getListed();
+    assert!(listed.len() == 1, "TEST: Adding duplicate increases size by 1");
+
+    c.add("192.168.0.0".to_string());
+    c.add("192.168.0.1".to_string());
+    c.add("192.168.0.2".to_string());
+    c.add("192.168.0.3".to_string());
+    c.add("192.168.0.0".to_string());
+    c.add("192.168.0.0".to_string());
+    let listed = c.getListed();
+    assert!(listed.len() == 4, "TEST: Adding additional adresses then duplicates fails");
+
+}
+fn remove_ip_test()
+{
+    println!("TEST: Channel - IP removal");
+    let config = channel::ChannelConfiguration::new(IPADDRESS.to_string(), 25565, "CUSTOM".to_string(), "fifo".to_string(), 20);
+    let mut c = channel::Channel::new(config);
+
+    c.add("192.168.0.1".to_string());
+    c.add("192.168.0.2".to_string());
+    c.add("192.168.0.3".to_string());
+    c.add("192.168.0.0".to_string());
+    c.add("192.168.0.0".to_string());
+    c.remove("192.168.0.0".to_string());
+    c.remove("192.168.0.2".to_string());
+
+    assert!(c.count() == 2, "TEST: Removing fail");
+
+    c.remove("192.168.0.1".to_string());
+    c.remove("192.168.0.3".to_string());
+
+    assert!(c.count() == 0, "TEST: Removing fail");
+
+    c.remove("192.168.0.1".to_string());
+    c.remove("192.168.0.3".to_string());
+
+    assert!(c.count() == 0, "TEST: Removing from empty fail");
+
 }
